In _get_freq_grid force repeats to an `int`
Appears that the number of repeats provided to Dask Array's `repeat`
must be of `int` exactly and not just some general integral type. We
have proposed a fix to Dask to relax this constraint in `repeat` to
accept all integral values, which has recently been accepted.

However it is quite peculiar that a Dask Array would have a shape that
is not just a `tuple` of `int`s. It appears there are some operations
with Dask Arrays that cause the shape to be a `tuple` of `numpy.int64`s
instead. This appears to be happening with filters from
`dask-ndfilters`. Though it is unclear whether the origin is there or in
Dask Array as well. In the interim, this should workaround the issue
until we better understand what is happening and resolve the underlying
cause.